+++ CHANGELOG.md
@@ -8,4 +8,5 @@ All notable changes to this project will be documented in this file.
   * [`PanBar`](./src/webgis4u/ol/control/PanBar.js)
   * [`LayerCheckbox`](./src/webgis4u/ol/control/LayerCheckbox.js)
   * [`LayerRadioButton`](./src/webgis4u/ol/control/LayerRadioButton.js)
+  * [`LegendDisplayable`](./src/webgis4u/ol/control/LegendDisplayable.js)
   * [`LegendVisibility`](./src/webgis4u/ol/control/LegendVisibility.js)